requirements.txt
# Core dependencies
Pillow>=10.0.0
PySide6>=6.5.0
numpy>=1.24.0

# Development dependencies
pytest>=7.4.0
pytest-qt>=4.2.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0

# Packaging
setuptools>=68.0.0
wheel>=0.41.0